+++ hadoop-ozone/ozonefs/src/main/java/org/apache/hadoop/fs/ozone/OzoneClientAdapter.java
@@ -27,7 +27,7 @@ import java.net.URI;
 import java.util.Iterator;
 
 /**
- * Lightweight adapter to separte hadoop/ozone classes.
+ * Lightweight adapter to separate hadoop/ozone classes.
  * <p>
  * This class contains only the bare minimum Ozone classes in the signature.
  * It could be loaded by a different classloader because only the objects in
