+++ hadoop-ozone/ozonefs/src/test/java/org/apache/hadoop/fs/ozone/TestOzoneFileInterfaces.java
@@ -139,7 +139,7 @@ public class TestOzoneFileInterfaces {
       fs = FileSystem.get(new URI(rootPath + "/test.txt"), conf);
     }
     o3fs = (OzoneFileSystem) fs;
-    statistics = o3fs.getOzoneFSOpsCountStatistics();
+    statistics = (OzoneFSStorageStatistics) o3fs.getOzoneFSOpsCountStatistics();
   }
 
   @After
